This is a combination of 12 commits.
author ehu <ehu@U008843.gi.ehu.es> 1568196849 +0200
committer Xabi Garde <xabier.gardeazabal@ehu.eus> 1592570161 +0200

This is a combination of 6 commits.tree a3ed7bb013007f6614a494445955f6b60c92e03a
author ehu <ehu@U008843.gi.ehu.es> 1568196849 +0200
committer Xabi Garde <xabier.gardeazabal@ehu.eus> 1592570061 +0200

This is a combination of 4 commits.tree e4bc1ca0df38041670fc7ffca5133d67cb25b390
author ehu <ehu@U008843.gi.ehu.es> 1568196849 +0200
committer Xabi Garde <xabier.gardeazabal@ehu.eus> 1592570027 +0200

This is a combination of 2 commits.
This is the 1st commit message:

This is a combination of 8 commits.

This is the 1st commit message:

Squash commit

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Author:    ehu <ehu@U008843.gi.ehu.es>
Date:      Wed Sep 11 12:14:09 2019 +0200

interactive rebase in progress; onto 267c049
Last commands done (40 commands done):
   reword 90666b8 Initial commit
   reword 90666b8 Initial commit
Next commands to do (22 remaining commands):
   squash 737ffef Reestruct fase1
   squash 3af92f0 Create README.md
You are currently editing a commit while rebasing branch 'xabi_changes' on '267c049'.

Initial commit

Changes to be committed:
      new file:   ControlRobot.cpp
      new file:   ControlRobot.h
      new file:   IRobotConnection.cpp
      new file:   IRobotConnection.h
      new file:   IRobotInstructionSet.cpp
      new file:   IRobotInstructionSet.h
      new file:   Instruction.cpp
      new file:   Instruction.h
      new file:   Makefile
      new file:   Serial.cpp
      new file:   Serial.h
      new file:   main.cpp
      new file:   nbproject/Makefile-Debug.mk
      new file:   nbproject/Makefile-Release.mk
      new file:   nbproject/Makefile-impl.mk
      new file:   nbproject/Makefile-variables.mk
      new file:   nbproject/Package-Debug.bash
      new file:   nbproject/Package-Release.bash
      new file:   nbproject/configurations.xml
      new file:   nbproject/private/Makefile-variables.mk
      new file:   nbproject/private/configurations.xml
      new file:   nbproject/private/downloads-10.141.78.221-root-22
      new file:   nbproject/private/downloads-192.168.1.204-root-22
      new file:   nbproject/private/downloads-192.168.1.221-root-22
      new file:   nbproject/private/downloads-192.168.43.161-root-22
      new file:   nbproject/private/launcher.properties
      new file:   nbproject/private/private.xml
      new file:   nbproject/private/timestamps-10.141.78.221-root-22
      new file:   nbproject/private/timestamps-192.168.1.204-root-22
      new file:   nbproject/private/timestamps-192.168.1.221-root-22
      new file:   nbproject/private/timestamps-192.168.43.161-root-22
      new file:   nbproject/project.xml

This is the commit message #2:

Reestruct fase1

This is the commit message #3:

Create README.md
This is the commit message #4:

minorchanges

This is the commit message #5:

Añadir Camara y Lidar

This is the commit message #6:

Añadir Componentes a Netbeans

This is the commit message #7:

asdffa

This is the commit message #8:

Test

This is the commit message #2:

asdas

This is the commit message #4:

Create .gitignore

This is the commit message #6:

Update main.cpp

This is the commit message #9:

Delete Makefile-Release.mk
This is the commit message #10:

Delete private.xml
This is the commit message #11:

This is the commit message #12:

Añadir OCVCam